refactor(signature): promote C14N to moxml top-level; drop XPath Filter
Addresses the architecture question raised in review: "Don't moxml
(or ../canon) already implement C14n?"

Right answer was to promote C14N out of the signature sub-namespace
into a top-level moxml feature, so it can serve any consumer (signature,
canon sibling project, future tooling) instead of being signature-only.

Moves:
- lib/moxml/signature/c14n.rb       -> lib/moxml/c14n.rb
- lib/moxml/signature/c14n/*.rb     -> lib/moxml/c14n/
- Moxml::Signature::C14n            -> Moxml::C14n (renamed)
- spec/moxml/signature/c14n/*       -> spec/moxml/c14n/
- lib/moxml.rb: new autoload :C14n entry

C14N is a core XML operation (parsing, serializing, XPath, C14N). Every
XML library has it. Living at the moxml top level makes the dependency
graph clean: canon (which already depends on moxml) can eventually
migrate its own ~2,400-line C14N implementation to use Moxml::C14n,
eliminating duplication without creating a circular dependency.

Removal:
- lib/moxml/signature/algorithms/xpath_filter_transform.rb — DELETED.
  The XPath Filter transform was a half-baked regex hack, not real XPath
  evaluation. The W3C spec marks XPath as RECOMMENDED, not REQUIRED;
  Best Practice 5 says "Try to avoid or limit XPath transforms"; Best
  Practice 22 prefers XPath Filter 2.0 over 1.0. The Enveloped Signature
  transform (the only transform most real signatures use) does not
  require XPath — it walks ancestors directly.

Documentation:
- TODO.complete/19-c14n-core-feature.md explains the move, the canon
  overlap, and the future adapter-delegated C14N opportunity
  (Nokogiri has native C14N via libxml2; other adapters would fall
  back to pure-Ruby).

96 specs, 0 failures, 0 rubocop offenses. Full existing suite (1896
examples, 174 pending) still passes.
