+++ lib/moxml/signature/algorithms.rb
@@ -73,7 +73,6 @@ module Moxml
           # Transforms
           Base64Transform
           EnvelopedSignatureTransform
-          XPathFilterTransform
 
           @builtins_loaded = true
         end
@@ -110,7 +109,7 @@ module Moxml
       autoload :EcdsaSha, "moxml/signature/algorithms/ecdsa_sha"
       autoload :DsaSha, "moxml/signature/algorithms/dsa_sha"
 
-      # Canonicalization (delegates to Moxml::Signature::C14n engine)
+      # Canonicalization (delegates to Moxml::C14n engine)
       autoload :ExcC14n10, "moxml/signature/algorithms/exc_c14n_10"
       autoload :InclusiveC14n10, "moxml/signature/algorithms/inclusive_c14n_10"
       autoload :InclusiveC14n11, "moxml/signature/algorithms/inclusive_c14n_11"
@@ -119,8 +118,6 @@ module Moxml
       autoload :Base64Transform, "moxml/signature/algorithms/base64_transform"
       autoload :EnvelopedSignatureTransform,
                "moxml/signature/algorithms/enveloped_signature_transform"
-      autoload :XPathFilterTransform,
-               "moxml/signature/algorithms/xpath_filter_transform"
     end
   end
 end
